Fix DoorProperties to use int (not uint) for JSON compat, improve error reporting
- ASP.NET JSON deserializer handles int? but may reject uint? from JSON integers
- Return 200 with error details on hardware endpoint instead of 400 to avoid
  swallowed error messages
- Cast to uint at the SDK call site

+++ genetec_sdk_service/src/GenetecSdkService.Api/Services/DoorService.cs
@@ -221,7 +221,7 @@ public class DoorService
                                 if (assignment.Hardware.HeldOpenTriggerTimeInSeconds.HasValue)
                                 {
                                     var triggerProp = held.GetType().GetProperty("TriggerTime");
-                                    triggerProp?.SetValue(held, assignment.Hardware.HeldOpenTriggerTimeInSeconds.Value);
+                                    triggerProp?.SetValue(held, (uint)assignment.Hardware.HeldOpenTriggerTimeInSeconds.Value);
                                 }
                             }
                         }
